Add test code for function pointers in checked block
The address-of (&) operator produces a checked array_ptr<T> type in a checked scope.   However, array_ptr<FunctionType> is not the correct type to produce when the address-of operator is applied to a function type.   It should produce a ptr type for a checked function pointer, not an array_ptr.