Use two files to test bounds checks for subscript operations.
Currently we are using 8 files to test bounds checks for subscript expressions.
Modify the two files for testing write operations involving subscript
expression to do all the tests, using pre-processing to change what operation
is tested.   We will rename these files and delete the other files in a
separate commit.