Unified test files for  dereferencing of pointer arithmetic.
Generalize the unified test file for subscript.c so that it can
be used to test the dereference operator applied to pointer
arithmetic.  We will delete the existing separate files for
testing dereferencing of pointer arithmetic in a separate check-in.

We use stub files deref-arith.c and deref-arith-opt.c, which compile
subscript.c with an #define so that the dereferencing and pointer arithmetic
is used in place of subscripting.