Split tests for bounds checking member expressions.
This improves the parallelism of running the Checked C regression tests.  The tests for bounds checking of member expressions were doing 8 different compiles: 4 for the dot operator and 4 for the arrow operators.  This change creates separate test automation driver files for the arrow operators.

The elapsed time for a test run on my 12 core Windows machine drops from 68 seconds to 45 seconds.  The original time before new tests were added was 30 seconds.  From back-of-the-envelope calculations, we would expect about a 50-60% increase in testing time with the additional tests.    This changes brings testing time into that range.

Testing:
- Passed automation testing on Linux
- Build and tested locally on Windows.