Stop generating LLVM IR as part of testing forany_parsing. (#177)
For_any parsing is generating LLVM IR files as part of test runs. This causes automation testing to fail because it creates the file in a source directory. The next automation run fails because it finds a file with no RUN line in it in the test directory. The fix is to not generate LLVM IR files.