Correct Interoperation type of stdlib.h's free(void*) (#190)
The important part of `free` is that you're passing an in-bounds pointer to an object to be free-ed. Given how we infer bounds for both array_ptrs and ptrs, require subsumption for function arguments, and that the smallest range possible for a pointer is `byte_count(1)`, this is the correct interface type rather than one that explicitly mentions array_ptr or ptr. 

We also change the interoperation type of the pointer parameter of `void *realloc(void*, size_t)` because it can also free memory. 

We add a set of tests that ensure these interface types for malloc, free, realloc and alloc_aligned operate with each other for every kind of pointer: unchecked, ptr and array_ptr.