Add variant of array_ptr for pointers to null-terminated arrays. (#192)
This change adds a new type `nt_array_ptr` that is a variant of `array_ptr` for pointers to null-terminated arrays.   The change:
- describes the new type in core-extensions.tex.   
- describes a new checked null-terminated array type go with `nt_array_ptr`
- describes the runtime bounds checking rules for nt_array_ptr.

For `nt_array_ptr` we allow inferred bounds if bounds are not declared.     Inference can widen bounds when it sees non-null memory access.   This raises the problem that memory accesses could depend on bounds that are not written down in the program.  The proposal is that any such accesses to memory must be provably in-bounds at compile time.

Most of the rules later in the specification for `array_ptr` apply to `nt_array_ptr`.    I initially started down the route of updating all occurrences of `array_ptr` to say "or `nt_array_ptr`".   I decided that was resulting in lots of changes that didn't help understanding much and made the text longer.   There is a paragraph in the description of `nt_array_ptr` that says that the rules for `array_ptr` usually apply to `nt_array_ptr` and that we mention `nt_array_ptr` only when the rules are different.

This change also cleans up some aspects of the spec:
- Remove the text that says bounds-checking is done for &e, if e is a dereference or subscript operation.  This is incorrect and doesn't match what the implementation does.
- Makes the description of bounds checks more precise, by describing the use of temporaries instead of expressions during bounds checking (the expression being bounds checked is not re-evaluated).
- Revise the section describing the deferral of the evaluation of bounds expression.

Still to do (in another pull request):
- Update the rules for conversions between checked pointer types.
- Describe the analysis that widens bounds for 'nt_array_ptr`.