Test bounds checking for variables with bounds-safe interfaces. (#200)
Variables and members with unchecked types may have bounds-safe interfaces.
When they are used in checked scopes, they should be treated as though they
have checked types.  This means that accesses to memory using pointers
stored in those variables or member should be bounds checked.

The changes adds tests of bounds checking of global and local variables with
bounds-safe interfaces and struct members with bounds-safe interfaces.  This
completes the work for issue #339.