Include null terminators in the number of elements for null-terminated arrays. (#210)
The Checked C specification proposed omitting the null terminator from the number of elements in a null-terminated array.   This would allow the count of elements to be identical for both the array and the bounds for an `nt_array_ptr`.  However, this is not how null-terminated string literals work in C.    String literals have an array type that includes the null terminator element in the size of the array.

To match existing C behavior, we now include the null terminator in the size of null-terminated checked arrays.  This updates the specification accordingly.