Replace bounds(none) with bounds(unknown) (#209)
The bounds expression bounds(none) has a potentially ambiguous English meaning.   The bounds could be unbounded or be unknown.  This replaces the bounds expression bounds(none) with bounds(unknown) in the specification.  This is mostly a search-and-replace of the boundsnone macro.  I also updated the text that defines the meaning of the bounds expression.