Test bounds checking for null-terminated pointers
Checked C allows reads at exactly the upper bound for null-terminated pointers
so that the potential terminating value can be examined.

This change adds new dynamic tests that check  that the the memory location at
the upper bound is accessible for null-terminated array pointers, and not
accessible for non-null-terminated array pointers.