Add examples of using null-terminated strings. (#219)
This adds examples of using null-terminated strings adapted from the "The C Programming Language, Second Edition" by Kernighan and Ritchie. The code illustrates some basic aspects of using strings, as well as the sometimes trickier aspects of writing to a string.

I plan to clean up the code once the compiler change for making count(0) be the default bounds for nt_array_ptr is complete.



I plan to clean up the code once the compiler change for making count(0) be the default bounds for nt_array_ptr is complete.