Use null-terminated array pointers in checked wrappers for C standard libraries (#229)
This change updates our checked wrappers for C standard libraries to use null-terminated array pointers for strings.
- I expanded the list of string functions and had to disable more optimized macros for GCC.
- In cases where I could make some parameters a function checked, but not other parameters checked,
I went ahead and did that.  This include functions such as `strcat` and `strncat`.   The fact that the destination buffer has an unchecked pointer type should be a warning to programmers trying to use them.
- For `strncat`, we have no way of expressing the bounds requirement in checked code, because the bounds requirement involves a function call.

Testing:
- Passed automated testing, including LNT testing.