Add error messages produced by bounds declaration checking (#237)
We are updating the compiler to do bounds declaration checking for constant-sized ranges (https://github.com/Microsoft/checkedc-clang/pull/414).   This causes new error messages to be produced for a number of tests.  For example, tests for typechecking might pass typechecking (as expected), but then fail bounds declaration checking..  

Update the tests with the expected error messages from clang (most of these errors were expected once enough bounds declaration checking came on-line and were flagged by TODO comments).

As an FYI, there are additional bounds declaration checking tests in the Checked C clang repo.  These contain detailed checks of the expected explanations, and are not placed in the Checked C repo.

The bounds declaration checking detects and gives errors for the bogus C-style casts described in issue #232.