Update specification.  Handle empty range special case.
Update the description of bounds checking in the specification to
allow allow writing a 0 at the upper bound of a pointer to a
null-terminated array.   Handle the special case where the allowed
memory range is empty (has no elements) even including the first
element of the null-terminated sequence in the range.

Add tests of this special case to nullterm_pointers.c.  This case
can arise when pointer arithmetic on the lower bound cause it
cross the upper bound (we always narrowing a lower bound by
adding a positive offset to it).