Remove -fcheckedc-extension flag. (#244)
The -fcheckedc-extension flag is now set by default for C files in the Checked C version of clang.  Remove the flag from the command lines for the test harness.

Also update a few error messages that changed after fixing bugs in the compiler.  The bugs were found after building the clang test suite with -fcheckedc-extension on by default.