Checked C typing rules for literals.  Also describe handling of initializers. (#249)
Describe extensions to C typing rules that allow literals to be used in checked scopes or where expressions with checked pointers are expected.   This completes the work for issue #207.   For checked scopes, string literals are given the type "null-terminated checked array of character", where "character" is the character type for the string literal.   For contexts where an expression with a checked pointer type is expected but not found, convert string literals and compound array literals to checked pointer types.

Also describe the rules for initializers.   The types of initializer values are determined by the  type of the variable being declared, so there isn't any change needed for Checked C.  

Add examples.

Move the section describing checked scopes earlier in Chapter 2, so that it precedes the discussion of literals (this is most of the change diff shown on GitHub).