Update rules for implicit conversions. (#251)
This updates the type checking rules for implicit casts. This addresses issues 
#150 and #206 n the specification.  It removes the subsections for each kind
of cast between pointer kinds and presents a unified framework instead.
- Use a diagram to describe what kinds of pointers can be converted to other
  kinds of pointers.  It is easier to visualize this than to explain in text.
  The conversions form a lattice. We can use this fact to more simply
  describe conversion rules.
- Allow conversions from checked void pointers.  We allow this by necessity
  already in the implementation.  It is needed so that malloc
  can be used in checked code.
- Mention that we plan to replace casts from checked void pointers.
- Move the description of assignment compatibility earlier.

In the description of C-style casts, remove the special case for ptr<void>.
It would be confusing to expand it now that we allow conversions from
array_ptr<void too. State in text instead that the conversions between
checked void pointers are allowed.

Add a description of the type checking rules for conditional expressions.
This was only partly covered, and that description was incomplete.
It is easier to explain this more generally.  Note that the compiler doesn't
quite match the spec.  It doen't allow casts between ptr<T> and array_ptr<T>,
which are now allowed.  I opened Checked C clang issue 444 to cover this.

Update some tests for type checking of conditional expressions to more
precisely check for arm types of the form 'T *' and "checked pointer to void"
that we're not merging them to T *.  This could then be converted to checked
pointer type.  The existing tests would pass and things would look OK, even
though from a typing perspective, it wouldn't be.