Update description of initialization requirements (#252)
Variables with checked pointer type that may be used to access memory must be initialized when they are declared.  This also applies to variables with members or elements with checked pointer type that may be used to access memory.  Update the specification to make this clear.   Include some examples.  This addresses issue #98.

The compiler already checks these requirements for scalar variables.  I opened https://github.com/Microsoft/checkedc-clang/issues/445 to track checking these requirements for struct and array variables.

Add tests for the struct/array cases to the Checked C tests.