Add dynamic tests of reads/writes via pointers with null-terminated interface types. (#259)
Add runtime testing of reads and writes through pointers with null-terminated interface types.  A pointer with a null-terminated type with no bounds information has a bounds of count(0).  A program can read the element at the pointer and write null.   Other reads/writes will cause bounds checking failures.

I extended bounds-safe-interfaces.c, which runs a sequence of read/tests, expecting conditions to be set properly for the nth test in the sequence to fail.   For bounds-safe interfaces for null-terminated types, we currently have no way to specify a length.  In this case, the bounds are fixed.  Instead, we vary the position where the read/write occurs to force a failure.