Checking of bounds declarations everywhere, expression equality facts. (#273)
This matches a compiler change that turns on checking of bounds declarations everywhere and teaches the compiler about simple equality facts.

Checking of bounds declarations is a static analysis that checks that declared bounds are valid (follow from existing bounds and other information in the environment). It was only on by default for checked scopes.  It is now on by default for unchecked scopes too.   The checking of bounds declarations is still simple, so it may produce unnecessary warnings.   It can be turned off using  `-Wno-check-bounds-decls` (everywhere), `-Wno-check-bounds-decl-unchecked-scope` (for unchecked scopes), and `-Wno-check-bounds-decls-checked-scope` (for checked scopes), 

  - Added more tests of expression equivalence (lexical_equality.c should be renamed because it now tests for expression equivalence using a few additional facts about C expressions).
  - Updated tests to handle the fact that checking of bounds declaration is on everywhere by default.  Replace a bunch of TODO's with expected warning messages. Also deleted some warnings that are no longer issued because the compiler can prove the bounds are valid.
  - Kept checking of bounds declarations on for most tests and only disabled it for some parsing and typechecking tests.
  - In some cases, made simple changes to avoid warnings about bounds declarations (the original cases will be captured in the new Checked C clang repo file test/CheckedC/static-checking/bounds-decl-challenges.c)