Test disallowing inferred bounds expressions with modifying expressions. (#278)
The current specification rules for inferring bounds for an expression can lead to inferred bounds that contain modifying expression.   These aren't allowed.  Add tests that check that the compiler is flagging them  (the Checked C version of clang suggests that the programmer use a temporary variable when this happens).

The particular problematic case that is likely to arise in practice is for member access: `e1.f` and `e1->f`.  If e1 contains a side-effect, we can't expression that bounds for `f` when other members are involved:
```
struct S {
  array_ptr<int> p : count(len);
  int len;
};
```
`e1 ->p` `has `bounds(e1->p, e1->p + e1->len)`.  This is problematic if `e1` has side-effects.   Note that the case where only `e1->p` is involved isn't a problem because we can use `current_expr_value`.