Update specification to match implementation.
I plan to release an updated version of the spec because the BOUNDS_CHECKED
pragma has been renamed to CHECKED_SCOPE.  Address some other issues where
the specification doesn't match the implementation.
- Address issue #100 (update motivation and definitions of
explicit dynamic check).  Add a statement that for dynamic_check(e),
e must be nonmodifying.
- Address issue #269.  Update specification to allow an interop
  type and a bounds expression as part of a bounds-safe interface.
- Change the wording to be clear that bounds-safe interface can be applied
  to unchecked arrays too.  This addresses some feedback in issue #272.
- Remove the requirement that every parameter for a function or
every member of a struct/member must be annotated with a bounds-safe interface,
if one of them is annotated.  The checking as implemented actually works on a
per parameter/per member basis.  Requiring that everything be converted
could get in the way of incremental conversion.