Fixup install of checked headers in subfolders (#316)
Fixes installation issue with checked headers in subfolders.

Checked headers not in the top-level directory in the source (e.g.
arpa/inet_checked.h) were being copied to the top-level directory
in the installation. This fixes that issue.
Specifically, the directory structure was sucessfully preserved when
copying the checked headers into the build directory, but then
subfolders were being removed from the path when copying from the
build directory into the install directory. I am still not sure
exactly why, but the documentation is clear that using DIRECTORY
instead of FILES in the install will solve the problem and preserve
directory structure. Tested in Linux (sudo make install).