Fix failing Windows installer build. (#322)
We added some new include files for POSIX in subdirectories last week.  The
build system was changed to try to use cmake functionality for recursively
installing a directory and its subdirectories.  This caused the build
of the Windows isntaller to break.

The fix is to revert last week's change and install the new subdirectories
individually.  This is the approach used in clang\lib\Headers\CMakeLists.txt.

This addresses https://github.com/Microsoft/checkedc-clang/issues/562.

Testing:
- Locally built a Windows installer.