Implicit void pointer casts for conditional expressions.
Add tests of implicit void pointer casts for arms of conditional expressions
in checked scopes.  For pointers to data containing checked pointers, these are
allowed in bounds_only scopes and not allowed in memory-safe checked scopes.
For pointers to data not containing checked pointers, these are allowed in both
kinds of scopes.