Test bounds checking of subscripting/dereferencing a pointer returned by a function. (#341)
A programmer can subscript a pointer returned by a function call or use a pointer dereference:

array_ptr<int> f(int len) : count(len);
int a = f(5)[3];
int b = *(f(5));

The subscripting or dereference operations in these cases should be bounds checked at runtime when the pointer is an array_ptr.  This change adds runtime tests of these cases, ensuring that bounds checking passes or fails as expected.  It adds tests for the following cases:
- the return bounds are constant.
- the return bounds depend on an argument.
- the return pointer is a null-terminated and has constant bounds.
- the return pointer is null-terminated and has bounds that depend on an argument.
- the return pointer is a multi-dimensional array where the outer dimension has variable-sized bounds.  C doesn't allow you to return an array type, so the return type in the last case is a pointer to an array.  

It also add tests that bounds checking works as expected for calls in checked scopes to functions with return bounds-safe interfaces.

We test uses of the subscripted or dereferenced value for several operations: reading a value from memory, writing a value to memory, post/pre-increment, and compound assignment.

We've adopted a different convention for organizing bounds checking tests than before.  Before, we would run a sequence of tests based on the arguments.  That meant that we had to supply a series of passing arguments to get to a failing test.   It was also hard to follow as we added cases. Now we just pass strings on the command line to name specific cases.  For these tests in this change, one string describes the type of return bounds and a second string names the operation.

Testing:
Passed automated testing for Linux and Windows with the compiler change at https://github.com/Microsoft/checkedc-clang/pull/605.