Fix checkedc unit test after bounds widening (#400)
After we have established that the value at the upper bound of an nt_array_ptr
s is non-null, we widen its bounds by 1. So the new upper bound of s is
(curr upper bound + 1). So in order to test writing at the upper we need to
write to *(s+1) instead of *s.