Allow the memory block passed to `realloc` to be zero-size.
This is analogous to https://github.com/microsoft/checkedc/pull/298 for
`free` and presumably is safe for the same reason.

Quick test case:

  #include <string.h>
  #include <stdlib.h>

  void foo() {
    // https://github.com/microsoft/checkedc-clang/issues/614
    _Nt_array_ptr<char> p = (_Nt_array_ptr<char>) strdup("hello");
    realloc(p, 15);
  }

I thought this might help with thttpd, which frequently calls `realloc`
on _Nt_array_ptrs with no known minimum length. It made no difference
there yet, but it's still the right thing to do.