Change the existing conditional headers to raise an #error if the original header is not available on the system.
As agreed with Sulekha. Conditionalizing the installation via CMake
might be even better, but that would take some more thought, especially
if we still want to test threads_checked.h on systems that don't have
threads.h.

Note that for the wrapper headers (`foo.h` as opposed to
`foo_checked.h`), we reverse the order of the conditionals because we
always want to use the `#error` message from `foo.h`, not the one from
`foo_checked.h`, which wouldn't make sense if the user originally
included `foo.h`.

The `#error` message is subject to change in Microsoft's code review.