Checked C macro test cases.
The Checked C compiler now supports an alternate syntax and macros for backward compatibility.  The technical details of the syntax and macros are documented at https://github.com/secure-sw-dev/checkedc/wiki/Proposed-extension-changes-to-improve-backward-compatibility.

This change modifies some of the tests in existing files to use the new syntax.  It only modifies a small fraction of the tests because of the syntactic nature of the changes.  This avoids duplicating test files, making it easier to modify tests later.