Clarify what features are actually supported.
This change updates the specification to make it clearer what
features are actually implemented in Checked C.
- The change groups the language extension features together and follows
them with a description of the static checking compilers should do.
- The existing specification interleaved ideas that had been
implemented with planned future work and speculative ideas.
Move the planned work to a new chapter at the end of the specification
and move most speculative ideas out of the specification entirely,
placing them in fragment.