Update Checked C specification to match the compiler implementation part 1 (#501)
The Checked C specification does not match the compiler implementation.   This is the first part of bringing the specification into alignment with the compiler implementation.  It addresses
- Move planned features to a new section in the specification.
- Remove speculative features from the specification.
- Move the description of static checking of bounds declarations to the end of the specification.
It also moves the related work section and the reject design to appendices.  They don't need to be in the specification and typically would be considered auxiliary material for any programming languages standard document.